Moved and renamed to WinEvents
