+++ src/main/java/com/stefancooper/SpigotUHC/Plugin.java
@@ -11,7 +11,7 @@ import com.stefancooper.SpigotUHC.events.BaseEvents;
 import com.stefancooper.SpigotUHC.events.ReviveEvents;
 import com.stefancooper.SpigotUHC.events.UHCLootEvents;
 import com.stefancooper.SpigotUHC.utils.UHCCommandTabCompleter;
-import com.stefancooper.SpigotUHC.utils.Win;
+import com.stefancooper.SpigotUHC.events.WinEvents;
 import org.bukkit.Bukkit;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
@@ -31,8 +31,8 @@ public class Plugin extends JavaPlugin implements Listener {
         Defaults.setDefaultGameRules(this.config);
         Bukkit.getPluginManager().registerEvents(new BaseEvents(config), this);
         Bukkit.getPluginManager().registerEvents(new ReviveEvents(config), this);
+        Bukkit.getPluginManager().registerEvents(new WinEvents(config), this);
         Bukkit.getPluginManager().registerEvents(new UHCLootEvents(config), this);
-        Bukkit.getPluginManager().registerEvents(new Win(config), this);
         started = false;
         System.out.println("UHC Plugin enabled");
         getCommand("uhc").setTabCompleter(new UHCCommandTabCompleter());
